requirements.txt
web3
requests
pandas
openpyxl
python-dotenv